+++ src/main/java/kz/bars/familybudget/repository/ExpenseCategoryRepo.java
@@ -5,10 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigInteger;
-
 @Repository
 @Transactional
-public interface ExpenseCategoryRepo extends JpaRepository<ExpenseCategory, BigInteger> {
+public interface ExpenseCategoryRepo extends JpaRepository<ExpenseCategory, Long> {
 
 }
